Build our own show parser
This replaces textfsm/ntc_templates, which are somewhat broken on Python 3.10

The refactoring this allowed us is a speedup, we no longer have to ask each
device for it's MAC table multiple times, we just grab the mac table, cdp table,
and etherchannel table one time each and parse them all at once.

There is improvement still, the etherchannel table is re-parsed each time we have to follow
a etherchannel link, but it is only requested from the device one time.
A future commit will address this.