Add script to put output into the order the input was
Ever since doing lookups in batches for speed, the output was not in the
same order as the input. So if you were copy pasting from a spreadsheet
for example, it was a lot of work to copy paste back. This script puts
things back in the right order, so you don't have to. It can be used
on the shell with a pipe from stdin (--stdin option) or from a file
after the fact.